more-wip

+++ AK/Types.h
@@ -51,7 +51,7 @@ using __ptrdiff_t = __PTRDIFF_TYPE__;
 #    endif
 #    if defined(AK_OS_WINDOWS)
 using ssize_t = MakeSigned<size_t>;
-using mode_t = int;
+using mode_t = unsigned short;
 #    endif
 
 #endif
